Root-mean-square error loss function ignoring bias

Subtracts the mean from the observed residuals before computing their root-mean-square error. This can be helpful when predicting time series where a bias due to a trend component is expected and can be captured by the innovation function choice.

Usage

loss_rmse_ignoring_bias(y_hat, y, ...)

Arguments

y_hat

A numeric vector representing predictions

y

A numeric vector representing observations

...

Additional arguments passed from other functions; ignored

Value

A scalar value

See also