Postprocess function making forecast samples non-negative while keeping the mean at the same value as it was before through scaling of the samples (unless the sample mean was originally negative).

While the identical sample mean can be enforced, this will generally decrease the sample variance as this is the only way to maintain the mean while turning negative samples to zero to enforce the non-negativity constraint.

Usage

non_negative_with_identical_mean(x)

Arguments

x

A numeric matrix of dimensions (k, 1)

Value

A numeric matrix of dimensions (k, 1)

Details

The returned samples will always be non-negative. The sample mean will not be identical if it cannot be achieved given the non-negative samples.

See also

Examples

non_negative_with_identical_mean(x = matrix(c(-1, 0, 1, 2), ncol = 1))
#>           [,1]
#> [1,] 0.0000000
#> [2,] 0.0000000
#> [3,] 0.6666667
#> [4,] 1.3333333

x <- matrix(rnorm(100, 0.5), ncol = 1)
summary(x)
#>        V1         
#>  Min.   :-2.2331  
#>  1st Qu.:-0.2726  
#>  Median : 0.5977  
#>  Mean   : 0.4819  
#>  3rd Qu.: 1.1883  
#>  Max.   : 2.5192  
summary(non_negative_with_identical_mean(x))
#>        V1        
#>  Min.   :0.0000  
#>  1st Qu.:0.0000  
#>  Median :0.3957  
#>  Mean   :0.4819  
#>  3rd Qu.:0.7867  
#>  Max.   :1.6679  

x <- matrix(rnorm(100, -1), ncol = 1)
summary(x)
#>        V1         
#>  Min.   :-3.2501  
#>  1st Qu.:-1.5320  
#>  Median :-0.9646  
#>  Mean   :-0.9347  
#>  3rd Qu.:-0.2403  
#>  Max.   : 1.3444  
# can't keep identical mean as original mean is negative
summary(non_negative_with_identical_mean(x))
#>        V1         
#>  Min.   :0.00000  
#>  1st Qu.:0.00000  
#>  Median :0.00000  
#>  Mean   :0.06918  
#>  3rd Qu.:0.00000  
#>  Max.   :1.34440